+++ packages/plugins/rsbuild-plugin-extension-manifest/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @addfox/rsbuild-plugin-extension-manifest
 
+## 0.1.1-beta.9
+
+### Patch Changes
+
+- @addfox/core@0.1.1-beta.9
+
 ## 0.1.1-beta.8
 
 ### Patch Changes